--- client/src/components/app.css
@@ -1,6 +1,0 @@
-.app {
-    text-size-adjust: none;
-    font-family: helvetica, arial, sans-serif;
-    line-height: 200%;
-    padding: 6px 20px 30px;
-}
