+++ client/src/index.jsx
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom';
 import { AppContainer } from 'react-hot-loader';
 // AppContainer is a necessary wrapper component for HMR
 
-import App from './components/app.jsx';
+import App from './react/app.jsx';
 
 const render = (Component) => {
     ReactDOM.render(
@@ -19,7 +19,7 @@ render(App);
 
 // Hot Module Replacement API
 if (module.hot) {
-    module.hot.accept('./components/app.jsx', () => {
+    module.hot.accept('./react/app.jsx', () => {
         render(App)
     });
 }
