+++ client/src/react/app.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import styles from './app.css';
+import styles from '../style/app.css';
 
 const App = () => (
 <div className={styles.app}>
-<h2>Hello, </h2>
+<h2>Welcome</h2>
 </div>
 );
