+++ client/src/style/app.css
@@ -1,0 +1,4 @@
+.app {
+    color: #c3c3c3;
+    background-color: #082141;
+}
